precision stuff

+++ models.py
@@ -18,13 +18,13 @@ MODELS = [
 
 WEIGHTS = {
     "EleutherAI/pythia-70m": torch.bfloat16,
-    "gpt2": torch.float16,
+    "gpt2": torch.float32,
     "EleutherAI/pythia-160m": torch.bfloat16,
-    "gpt2-medium": torch.float16,
+    "gpt2-medium": torch.float32,
     "EleutherAI/pythia-410m": torch.bfloat16,
-    "gpt2-large": torch.float16,
+    "gpt2-large": torch.float32,
     "EleutherAI/pythia-1b": torch.bfloat16,
-    "gpt2-xl": torch.float16,
+    "gpt2-xl": torch.float32,
     "EleutherAI/pythia-1.4b": torch.bfloat16,
     "EleutherAI/pythia-2.8b": torch.bfloat16,
     "sharpbai/alpaca-7b-merged": torch.bfloat16,
